Trabalhos/Pizzaria_SGAV/requirements.txt
greenlet==3.2.2
PyMySQL==1.1.1
pyodbc==5.2.0
python-dotenv==1.0.1
SQLAlchemy==2.0.41
typing_extensions==4.13.2
